Rename instrumentation tests
